Added functionality for pruning mgo/txn's transactions collection
If left alone, the collection used by mgo/txn to coordinate and track
client-side transactions will grow without bound.

This change extends the Runner interface with a PruneTransactions
method. PruneTransactions can be called at any time to remove
mgo/txn's documents for applied and aborted transactions. It is
intended to be called periodically.