Only prune txns when it would help
Pruning is a heavyweight operation so pruning will now only be
attempted if there has been significant growth in txns collection since
the last pruning run. To support this, the transaction count after the
last pruning attempt is now recorded in a txns.prune collection.

PruneTransactions has now become MaybePruneTransactions.

This should greatly reduce the amount of unnecessary pruning attempts
and therefore the impact on MongoDB performance.

Also: removed the use of juju/errors which was introduced with the
previous pruning work.