Record historical pruning stats
Instead of just recording details of the last pruning run in the
txns.prune collection, details for all pruning runs are now kept. More
stats are kept now to allow analysis of pruning performance. For quick
access, a pointer document is used to track the most recent result
(required for regular pruning checks).