Create a type for cleaning up txn-queues.
This is originally focused on cleaning up txns.stash, but can be used for
other collections as well. Integrate the stash cleanup into our normal
'purgetxn' operations. This iterates through the documents batching up
transaction tokens and checking to see which of these tokens represent
transactions that are already marked as applied or aborted. We can then remove
all of those references from the documents. If we are updating a txn.stash
document, we can then remove the document if all of the transactions it was
involved in have been completed.