Enable cleaning up all collections.
This requires a bunch of changes to the test suite, as most of the code
assumed we would still have the last transaction touching a document left
behind after pruning, but now that's no longer the case.

Include a helper function to make it easier to set up a pending transaction
that shouldn't be removed.

Make the logic of when to prune slightly more complex.  Instead of just a
'pruneFactor', we now can take into account a minimum number of transactions
we want to see, so that we don't fast-loop on pruning everytime any
transaction shows up.

Skip loading documents that don't have any txn-queue entries.