avoid using arbitrarily large $in request
The DBOracle.CompletedTokens method uses
a $in request, which can be very slow
when there are a large number of tokens.
By using a map instead, runtime can be significantly
reduced (from >30 minutes to 3 seconds in
some cases).

Also trace errors, making it easier to see where
the errors come from.