Expose the batch size and number of batches.
We go ahead and do the loop inside MaybePruneTransactions, rather than
exposing it entirely. Otherwise we would have needed to expose the
CleanupStats. This should allow us to update github.com/juju/juju with
the configuration knobs as to how we want to batch our transactions
(default 1M, 100 max batches).