Lots of refactoring, a couple small algorithm tweaks.
Break things up into multiple functions, hopefully making it a lot
easier to read what is going on.

Change the docWithQueue so we cache the transformation from token to txn
ids. Since we're likely passing over the same documents over and over,
it makes sense to not do the conversion millions of times.

Change the 'find tokens to pull', to bias towards not having any
allocations when there is nothing to do, which is the case 99% of the
time.