Major refactor.
Move the ability to handle multiple processing routines into the
IncrementalPruner so that we can keep the stats all together in a single
struct, rather than multiple structs.

Move WaitGroup so it gets added when we spawn a goroutine as it should
have been, rather than side-effects of sending a message on the channel
to indicate that we are done.

Still needs live testing.